git-rebase: Support operating on region of commit lines
Teach git-rebase-set-action to act on the lines in an active region.
This affects the set of commands that target commit lines, as well as
git-rebase-kill-line, which targets rebase lines of all action types.

For other commands that don't operate on commit lines---such as
"exec", "noop", and "merge"---it's not clear that there's any sensible
behavior for an active region, so leave those be.

Closes #4172.