git-rebase-set-action: Absorb "kill line" logic
Commands that operate on a commit line and change the action type
(pick, edit, fixup, ...) all go through git-rebase-set-action.  On the
other hand, git-rebase-kill-line doesn't go through this helper
because its logic diverges a good amount.

However, the set-action-based commands and git-rebase-kill-line will
be reworked to support operating on a region, at which point
git-rebase-kill-line will end up sharing quite a bit of code with the
other commands.  To avoid duplicating the region handling, move the
custom kill logic into git-rebase-current-line as a special case.